trdi_python_requirements.txt
appdirs==1.4.3
argcomplete==1.9.4
asn1crypto==0.24.0
astroid==2.0.4
autopep8==1.4.3
bcrypt==3.1.5
beautifulsoup4==4.6.3
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
Django==2.1.4
escpos==1.5
Flask==1.0.2
future==0.17.1
idna==2.7
image==1.5.27
iso8601==0.1.12
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
libusb==1.0.22b2
libusb1==1.7
MarkupSafe==1.1.0
mccabe==0.6.1
paramiko==2.4.2
Pillow==5.3.0
praw==6.0.0
prawcore==1.0.0
pyasn1==0.4.4
pycodestyle==2.4.0
pycparser==2.19
pylint==2.1.1
PyNaCl==1.3.0
pyserial==3.4
python-dateutil==2.7.5
python-escpos==2.2.0
python-printer-escpos==0.0.3
pytz==2018.7
pyusb==1.0.2
PyYAML==3.13
qrcode==6.0
requests==2.21.0
serial==0.0.87
six==1.11.0
update-checker==0.16
urllib3==1.24.1
viivakoodi==0.8.0
virtualenv==16.1.0
virtualenv-clone==0.4.0
watson-developer-cloud==2.4.4
websocket-client==0.47.0
Werkzeug==0.14.1
wrapt==1.10.11
